Created bemail.php
A class that contains a recursive function that takes a given date returns the next date that is not either a weekend or a holiday. It also has a test to see whether it's after 1pm on the current day, but that test is commented out because either the controller or the view of this MVC is doing that job somewhere else.