Create admin_new_field.php
When a new custom form field was created, if the "name" field of the new form form (not a typo) wasn't filled in, the form would submit and claim success, but the field did not save. 

Normally in CakePHP the solution would be to add a validation rule to that field. Such a rule would be placed in plugins/EmailForms/Model/EmailForms.php, in the $validate array. 

However in this case, EmailForms uses a helper from another plugin, CustomFields. The validation rule would have to go in the CustomFields model. However, that rule would apply to every Name field of every CustomField throughout our CMS, and we only want it to apply to EmailForms CustomField name fields. 

The simplest solution was to validate the form in jQuery, disable the Submit button, and throw up an error if someone tried to submit the form without filling that particular field. 

Since I don't know how to intercept a form in jQuery after Submit has been pressed, I triggered the Mouseover event for the div that contains 2 Submit buttons (save&continue editing, save&return to index). 

The biggest challenge  for me was finding the selector that would grab all the name input fields on the form, and nothing else. Reddit helped. It was $('input[name$="[name]"]'). 

This is my first professional use of jQuery.